@@ -1,7 +1,7 @@
 {
-  "name": "@cognite/reveal-streaming-examples",
+  "name": "@cognite/reveal-examples",
   "version": "0.1.0",
-  "description": "Reveal streaming examples",
+  "description": "Reveal examples",
   "main": "dist/index.js",
   "types": "dist/src/index.d.ts",
   "files": [
@@ -19,7 +19,7 @@
   "author": "Svenn-Arne Dragly <dragly@cognite.com>",
   "contributors": [
     "Lars Moastuen <lars.moastuen@cognite.com>",
-    "Nils Petter Fremming <nils.fremming@cognite.com>"
+    "Joachim Bjørne <joachim.bjorne@cognite.com>"
   ],
   "license": "Apache-2.0",
   "glslify": {
@@ -29,48 +29,45 @@
   },
   "devDependencies": {
     "@cognite/tslint-config-cognite-3d": "^1.0.4",
-    "@tweenjs/tween.js": "^17.2.0",
-    "@types/cesium": "^1.59.3",
+    "@tweenjs/tween.js": "^18.5.0",
     "@types/dat.gui": "^0.7.5",
     "@types/gl-matrix": "^2.4.5",
-    "@types/jest": "^24.0.23",
-    "@types/node": "^12.12.11",
-    "@types/tween.js": "^17.2.0",
-    "@wasm-tool/wasm-pack-plugin": "^1.0.1",
-    "copy-webpack-plugin": "^5.1.0",
-    "css-loader": "^3.2.0",
-    "file-loader": "^5.0.2",
-    "html-webpack-plugin": "3.2.0",
+    "@types/jest": "^25.1.4",
+    "@types/node": "^13.9.4",
+    "@types/tween.js": "^18.5.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.2",
+    "file-loader": "^6.0.0",
+    "html-webpack-plugin": "4.0.2",
     "prettier": "^1.19.1",
-    "raw-loader": "^3.1.0",
-    "rimraf": "^3.0.0",
+    "raw-loader": "^4.0.0",
+    "rimraf": "^3.0.2",
     "source-map-loader": "^0.2.4",
-    "ts-loader": "6.2.1",
+    "ts-loader": "6.2.2",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-no-circular-imports": "^0.7.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "3.7.2",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
+    "tslint-plugin-prettier": "^2.3.0",
+    "typescript": "3.8.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
     "webpack-log": "^3.0.1",
     "webpack-node-externals": "^1.7.2",
-    "worker-plugin": "^3.2.0"
+    "worker-plugin": "^4.0.2"
   },
   "dependencies": {
-    "gl-matrix": "^3.1.0",
-    "camera-controls": "^1.15.0",
-    "cesium": "^1.64.0",
+    "@cognite/potree-core": "^1.1.1",
+    "@cognite/reveal": "../viewer/dist",
+    "@cognite/sdk": "^2.26.0",
+    "@xailabs/three-renderer-stats": "^1.0.4",
+    "camera-controls": "^1.18.3",
     "comlink": "^4.1.0",
     "dat.gui": "^0.7.6",
-    "@cognite/potree-core": "^1.0.9",
-    "@cognite/reveal": "../viewer",
-    "@cognite/sdk": "^2.13.0",
-    "@xailabs/three-renderer-stats": "^1.0.4",
-    "postprocessing": "^6.8.9",
-    "three": "0.110.0",
+    "gl-matrix": "^3.2.1",
     "glslify": "^7.0.0",
-    "glslify-import": "^3.1.0"
+    "postprocessing": "^6.13.2",
+    "rxjs": "^6.5.4",
+    "three": "^0.115.0"
   }
 }
