@@ -1,15 +1,15 @@
 {
   "name": "@cognite/reveal",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "Reveal streaming component",
-  "main": "dist/index.js",
-  "types": "dist/src/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
-    "release": "webpack --env.three-examples false --env.cesium-examples false",
-    "build": "webpack --env.development",
+    "release": "npm run mkdist && webpack",
+    "build": "npm run mkdist && webpack --env.development",
+    "mkdist": "run-script-os",
+    "mkdist:win32": "@powershell $(if(-not(Test-Path dist -PathType Container)) { mkdir dist;$? } else { $true;$? }) -and $(Copy-Item -Path package.json, README.md -Destination dist;$?)",
+    "mkdist:linux:darwin": "mkdir -p dist && cp package.json README.md dist",
     "clean": "rimraf -rf dist/ target/ pkg/",
     "test": "jest",
     "coverage": "jest --coverage",
@@ -19,9 +19,10 @@
   "author": "Svenn-Arne Dragly <dragly@cognite.com>",
   "contributors": [
     "Lars Moastuen <lars.moastuen@cognite.com>",
-    "Nils Petter Fremming <nils.fremming@cognite.com>"
+    "Joachim Bjørne <joachim.bjorne@cognite.com>"
   ],
   "license": "Apache-2.0",
+  "sideEffects": false,
   "glslify": {
     "transform": [
       "glslify-import"
@@ -29,44 +30,51 @@
   },
   "devDependencies": {
     "@cognite/tslint-config-cognite-3d": "^1.0.4",
-    "@types/cesium": "^1.59.3",
     "@types/gl-matrix": "^2.4.5",
-    "@types/jest": "^24.0.23",
-    "@types/node": "^12.12.11",
-    "@wasm-tool/wasm-pack-plugin": "^1.0.1",
-    "copy-webpack-plugin": "^5.1.0",
-    "file-loader": "^5.0.2",
+    "@types/jest": "^25.1.4",
+    "@types/node": "^13.9.4",
+    "@types/tween.js": "^18.5.0",
+    "@wasm-tool/wasm-pack-plugin": "^1.1.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "file-loader": "^6.0.0",
+    "gl": "^4.5.0",
     "glslify-loader": "^2.0.0",
-    "html-webpack-plugin": "3.2.0",
-    "jest": "^24.9.0",
+    "html-webpack-plugin": "4.0.2",
+    "jest": "^25.2.2",
     "jest-canvas-mock": "^2.2.0",
-    "jest-extended": "^0.11.2",
-    "jest-fetch-mock": "^2.1.2",
+    "jest-extended": "^0.11.5",
+    "nock": "^12.0.3",
     "prettier": "^1.19.1",
-    "raw-loader": "^3.1.0",
-    "rimraf": "^3.0.0",
-    "ts-jest": "^24.2.0",
-    "ts-loader": "6.2.1",
+    "raw-loader": "^4.0.0",
+    "rimraf": "^3.0.2",
+    "run-script-os": "^1.1.1",
+    "ts-jest": "^25.2.1",
+    "ts-loader": "6.2.2",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-no-circular-imports": "^0.7.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "3.7.2",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
+    "tslint-plugin-prettier": "^2.3.0",
+    "typescript": "3.8.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
     "webpack-log": "^3.0.1",
     "webpack-node-externals": "^1.7.2",
-    "worker-plugin": "^3.2.0"
+    "whatwg-fetch": "^3.0.0",
+    "worker-plugin": "^4.0.2"
   },
   "dependencies": {
-    "@cognite/potree-core": "^1.0.9",
-    "@cognite/sdk": "^2.13.0",
-    "cesium": "^1.63.1",
-    "three": "0.110.0",
-    "gl-matrix": "^3.1.0",
+    "@cognite/potree-core": "^1.1.1",
+    "@cognite/sdk": "^2.26.0",
+    "@cognite/three-combo-controls": "^1.4.1",
+    "@tweenjs/tween.js": "^18.5.0",
+    "@types/lodash": "^4.14.149",
+    "comlink": "^4.1.0",
+    "gl-matrix": "^3.2.1",
     "glslify": "^7.0.0",
     "glslify-import": "^3.1.0",
-    "comlink": "^4.1.0"
+    "lodash": "^4.17.15",
+    "rxjs": "^6.5.4",
+    "three": "^0.115.0"
   }
 }
