@@ -3,12 +3,9 @@ use std::io::{stdout, BufReader};
 use structopt::StructOpt;
 
 use serde_derive::{Deserialize, Serialize};
-use serde_json;
-use serde_yaml;
 
 #[derive(StructOpt)]
 struct Options {
-    root_file: String,
     sector_file: String,
     #[structopt(short, long)]
     yaml: bool,
@@ -31,21 +28,10 @@ enum Output {
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let options = Options::from_args();
 
-    let root_file = options.root_file;
     let sector_file = options.sector_file;
 
-    let root_reader = BufReader::new(File::open(root_file)?);
     let sector_reader = BufReader::new(File::open(sector_file)?);
-    let root_sector = i3df::parse_root_sector(root_reader)?;
-    let attributes = match root_sector.header.attributes {
-        Some(x) => x,
-        None => {
-            return Err(Box::new(i3df::error::Error::new(
-                "Attributes missing on root sector",
-            )))
-        }
-    };
-    let raw_sector = i3df::parse_sector(&attributes, sector_reader)?;
+    let raw_sector = i3df::parse_sector(sector_reader)?;
     let sector: Output = if options.renderables {
         Output::RenderableSector(Box::new(i3df::renderables::convert_sector(&raw_sector)))
     } else {
