@@ -2,7 +2,6 @@ use std::io::{self, BufRead, BufReader, Cursor, Read};
 
 use byteorder::{LittleEndian, ReadBytesExt};
 
-use nalgebra;
 use serde_derive::{Deserialize, Serialize};
 
 use wasm_bindgen::prelude::*;
@@ -61,8 +60,7 @@ pub fn parse_scene(reader: impl BufRead) -> Result<Scene, Error> {
     parse_scene_data(reader)
 }
 
-pub fn parse_root_sector(mut reader: impl BufRead) -> Result<Sector, Error> {
-    //let mut reader = BufReader::new(reader);
+pub fn parse_sector(mut reader: impl BufRead) -> Result<Sector, Error> {
     let size = reader.read_u32::<LittleEndian>()?;
 
     // read bytes
@@ -74,31 +72,8 @@ pub fn parse_root_sector(mut reader: impl BufRead) -> Result<Sector, Error> {
     let header = parse_sector_header(&mut input)?;
     let attributes = match &header.attributes {
         Some(x) => x,
-        None => return Err(error!("Attributes missing on root sector")),
+        None => return Err(error!("Attributes missing on sector")),
     };
-    let primitive_collections = generated::parse_primitives(&mut input, attributes)?;
-    //assert!(reader.eof()?);
-    Ok(Sector {
-        header,
-        primitive_collections,
-    })
-}
-
-pub fn parse_sector(
-    attributes: &SectorAttributes,
-    mut reader: impl BufRead,
-) -> Result<Sector, Error> {
-    //let mut reader = BufReader::new(reader);
-    // read number of bytes
-    let size = reader.read_u32::<LittleEndian>()?;
-
-    // read bytes
-    let mut sector = vec![0; size as usize].into_boxed_slice();
-    reader.read_exact(&mut sector)?;
-
-    // parse sector
-    let mut input = BufReader::new(Cursor::new(sector));
-    let header = parse_sector_header(&mut input)?;
     let primitive_collections = generated::parse_primitives(&mut input, attributes)?;
     Ok(Sector {
         header,
@@ -107,13 +82,7 @@ pub fn parse_sector(
 }
 
 pub fn parse_scene_data(mut reader: impl BufRead) -> Result<Scene, Error> {
-    //let mut reader = BufReader::new(reader);
-
-    let root_sector = parse_root_sector(&mut reader)?;
-    let attributes = match &root_sector.header.attributes {
-        Some(x) => x,
-        None => return Err(error!("Attributes missing on root sector")),
-    };
+    let root_sector = parse_sector(&mut reader)?;
     let root_sector_id = root_sector.header.sector_id as usize;
 
     let mut other_sectors = Vec::new();
@@ -123,7 +92,7 @@ pub fn parse_scene_data(mut reader: impl BufRead) -> Result<Scene, Error> {
             break;
         }
 
-        let sector = parse_sector(&attributes, &mut reader)?;
+        let sector = parse_sector(&mut reader)?;
 
         other_sectors.push(sector);
     }
@@ -147,7 +116,7 @@ pub fn parse_sector_header(mut input: &mut impl BufRead) -> Result<SectorHeader,
         ));
     }
     let format_version = input.read_u32::<LittleEndian>()?;
-    assert_eq!(format_version, 7);
+    assert_eq!(format_version, 8);
     let optimizer_version = input.read_u32::<LittleEndian>()?;
 
     let sector_id = input.read_u64::<LittleEndian>()? as usize;
@@ -299,7 +268,7 @@ pub fn decode_array_texture(mut reader: impl Read) -> Result<Vec<Texture>, Error
     let mut array = Vec::new();
 
     for _ in 0..item_count {
-        let file_id = reader.read_u64::<LittleEndian>()?;
+        let file_id = reader.read_u64::<LittleEndian>()? as f64;
         let width = reader.read_u16::<LittleEndian>()?;
         let height = reader.read_u16::<LittleEndian>()?;
         let _reserved = reader.read_u32::<LittleEndian>()?;
