@@ -3,17 +3,18 @@ module.exports = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
-  testRegex: '(/__tests__/.*(test|spec|Test).*|(\\.|/)(test|spec|Test))\\.tsx?$',
+  testRegex: '(/__tests__/.*\\.test\\..*|\\.(test|spec|Test))\\.tsx?$',
   moduleDirectories: ['node_modules'],
   moduleFileExtensions: ['ts', 'js', 'json', 'node'],
   moduleNameMapper: {
     '^@/(.*)': '<rootDir>/src/$1',
     '\\.(frag|vert)$': '<rootDir>/src/__mocks__/glslMocks.js',
   },
+  coverageDirectory: '../coverage',
   collectCoverageFrom: [
-    "src/",
     "!src/__tests__/**/*.ts",
-    "!src/examples/**/*.ts",
+    "!**/*.d.ts",
+    "!**/*.json"
   ],
   automock: false,
   setupFiles: [
