@@ -1,8 +1,8 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
-import { traverseDepthFirst } from '../../utils/traversal';
+import { traverseDepthFirst, traverseUpwards } from '../../utils/traversal';
 
 describe('traversal', () => {
   type Element = {
@@ -69,5 +69,34 @@ describe('traversal', () => {
 
     // Assert
     expect(visitedOrder.join(',')).toBe([0, 1, 2, 3, 6, 7].join(','));
+  });
+
+  interface WithParentNode {
+    readonly id: number;
+    readonly parent?: WithParentNode;
+  }
+
+  test('traverseUpwards', () => {
+    // Arrange
+    const leaf: WithParentNode = {
+      id: 5,
+      parent: {
+        id: 4,
+        parent: {
+          id: 3
+        } as WithParentNode
+      } as WithParentNode
+    };
+    const visitedOrder: number[] = [];
+    const visitor = (element: WithParentNode) => {
+      visitedOrder.push(element.id);
+      return element.id !== 3;
+    };
+
+    // Act
+    traverseUpwards(leaf, visitor);
+
+    // Assert
+    expect(visitedOrder.join(',')).toBe([5, 4, 3].join(','));
   });
 });
