@@ -1,13 +1,13 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 import * as THREE from 'three';
 import { mat4 } from 'gl-matrix';
-import { PointCloudModel } from '../../../../datasources/PointCloudModel';
+import { PointCloudModel } from '../../../../models/pointclouds/PointCloudModel';
 import { createThreeJsPointCloudNode } from '../../../../views/threejs';
 import { FetchPointCloudDelegate } from '../../../../models/pointclouds/delegates';
-import { SectorModelTransformation } from '../../../../models/sector/types';
+import { SectorModelTransformation } from '../../../../models/cad/types';
 import { PotreeGroupWrapper } from '../../../../views/threejs/pointcloud/PotreeGroupWrapper';
 
 describe('createThreeJsPointCloudNode', () => {
