@@ -1,10 +1,10 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 import { Versioned3DFile } from '@cognite/sdk';
 
-export const supportedVersions = [1, 2, 3, 7];
+export const supportedVersions = [8];
 
 export function getNewestVersionedFile(files: Versioned3DFile[]): Versioned3DFile {
   return files
