@@ -1,12 +1,13 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 import { mat4 } from 'gl-matrix';
-import { PointCloudModel } from '../../PointCloudModel';
 import { FetchPointCloudDelegate } from '../../../models/pointclouds/delegates';
-import { SectorModelTransformation } from '../../../models/sector/types';
+import { SectorModelTransformation } from '../../../models/cad/types';
 import { PointCloudLoader } from '../../../utils/potree/PointCloudLoader';
+import { EptLoader } from '../../../utils/potree/EptLoader';
+import { PointCloudModel } from '../../../models/pointclouds/PointCloudModel';
 
 const identity = mat4.identity(mat4.create());
 
@@ -14,9 +15,16 @@ export function createLocalPointCloudModel(url: string): PointCloudModel {
   const fetchPointCloud: FetchPointCloudDelegate = async () => {
     const transform: SectorModelTransformation = {
       modelMatrix: identity,
-      inverseModelMatrix: mat4.invert(mat4.create(), identity)!
+      inverseModelMatrix: identity
     };
-    return [await PointCloudLoader.load(url), transform];
+
+    if (url.endsWith('ept.json')) {
+      // Entwine format
+      return [await EptLoader.load(url), transform];
+    } else {
+      // Potree format
+      return [await PointCloudLoader.load(url), transform];
+    }
   };
   return [fetchPointCloud];
 }
