@@ -1,9 +1,9 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 // TODO potentially private types
-export { FetchSectorMetadataDelegate, FetchSectorDelegate, FetchCtmDelegate } from './models/sector/delegates';
+export { FetchSectorDelegate, FetchCtmDelegate } from './models/cad/delegates';
 
 // TODO move these types into potree-core as TypeScript definitions and remove them from here
 export { PotreePointColorType, PotreePointShape } from './views/threejs/pointcloud/enums';
@@ -11,3 +11,11 @@ export { PotreeNodeWrapper } from './views/threejs/pointcloud/PotreeNodeWrapper'
 export { PotreeGroupWrapper } from './views/threejs/pointcloud/PotreeGroupWrapper';
 
 export { suggestCameraConfig } from './utils/cameraUtils';
+export { traverseDepthFirst, traverseUpwards } from './utils/traversal';
+export { determineSectorsByProximity, DetermineSectorsByProximityInput } from './models/cad/determineSectors';
+export { WantedSector } from './data/model/WantedSector';
+export { LevelOfDetail } from './data/model/LevelOfDetail';
+
+export { SectorCuller } from './culling/SectorCuller';
+export { ProximitySectorCuller } from './culling/ProximitySectorCuller';
+export { ByVisibilityGpuSectorCuller } from './culling/ByVisibilityGpuSectorCuller';
