@@ -1,23 +1,28 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 // @ts-ignore
 import * as Potree from '@cognite/potree-core';
 import { PotreePointSizeType, PotreePointColorType, PotreePointShape } from './enums';
+import { fromThreeJsBox3 } from '../utilities';
+import { Box3 } from '../../../utils/Box3';
+import { vec3 } from 'gl-matrix';
 
 /**
  * Wrapper around `Potree.PointCloudOctree` with some convinence functions.
  */
 export class PotreeNodeWrapper {
-  readonly octtree: Potree.PointCloudOctree;
+  readonly octtree: Potree.PointCloudOctreeNode;
 
-  constructor(octtree: Potree.PointCloudOctree) {
+  constructor(octtree: Potree.PointCloudOctreeNode) {
     this.octtree = octtree;
     this.pointSize = 2;
     this.pointSizeType = PotreePointSizeType.Adaptive;
     this.pointColorType = PotreePointColorType.Rgb;
     this.pointShape = PotreePointShape.Circle;
+
+    this.pointBudget = 2_000_000;
   }
 
   get pointSize(): number {
@@ -32,6 +37,25 @@ export class PotreeNodeWrapper {
   }
   set pointSizeType(type: PotreePointSizeType) {
     this.octtree.material.pointSizeType = type;
+  }
+
+  get pointBudget(): number {
+    return this.octtree.pointBudget;
+  }
+  set pointBudget(count: number) {
+    this.octtree.pointBudget = count;
+  }
+  get visiblePointCount(): number {
+    return this.octtree.numVisiblePoints || 0;
+  }
+
+  get boundingBox(): Box3 {
+    const bbox: THREE.Box3 = this.octtree.root.tightBoundingBox;
+    const box = fromThreeJsBox3(bbox);
+    // Apply transformation to switch axes
+    const min = vec3.fromValues(box.min[0], box.min[2], -box.min[1]);
+    const max = vec3.fromValues(box.max[0], box.max[2], -box.max[1]);
+    return new Box3([min, max]);
   }
 
   get pointColorType(): PotreePointColorType {
