@@ -1,11 +1,11 @@
 /*!
- * Copyright 2019 Cognite AS
+ * Copyright 2020 Cognite AS
  */
 
 import * as THREE from 'three';
 import { Box3 } from '../../utils/Box3';
-import { vec3, mat4, vec4 } from 'gl-matrix';
-import { SectorModelTransformation } from '../../models/sector/types';
+import { vec3, mat4 } from 'gl-matrix';
+import { SectorModelTransformation } from '../../models/cad/types';
 
 export function fitCameraToBoundingBox(camera: THREE.Camera, bounds: Box3, radiusFactor: number = 4) {
   const min = new THREE.Vector3(bounds.min[0], bounds.min[1], bounds.min[2]);
@@ -30,22 +30,23 @@ export function fitCameraToBoundingBox(camera: THREE.Camera, bounds: Box3, radiu
   camera.position.set(position.x, position.y, position.z);
 }
 
-export function toThreeJsBox3(box: Box3): THREE.Box3 {
-  return new THREE.Box3(toThreeVector3(box.min), toThreeVector3(box.max));
-}
-
 const toThreeVector3Vars = {
   result: vec3.create()
 };
 
-// TODO add out parameter
-export function toThreeVector3(v: vec3, modelTransformation?: SectorModelTransformation): THREE.Vector3 {
+export function toThreeVector3(
+  out: THREE.Vector3,
+  v: vec3,
+  modelTransformation?: SectorModelTransformation
+): THREE.Vector3 {
   if (!modelTransformation) {
-    return new THREE.Vector3(v[0], v[1], v[2]);
+    out.set(v[0], v[1], v[2]);
+    return out;
   }
   const { result } = toThreeVector3Vars;
   vec3.transformMat4(result, v, modelTransformation.modelMatrix);
-  return new THREE.Vector3(result[0], result[1], result[2]);
+  out.set(result[0], result[1], result[2]);
+  return out;
 }
 
 const toThreeMatrix4Vars = {
@@ -92,4 +93,27 @@ export function fromThreeMatrix(out: mat4, m: THREE.Matrix4, modelTransformation
     return mat4.multiply(out, modelTransformation.inverseModelMatrix, out);
   }
   return out;
+}
+
+const toThreeJsBox3Vars = {
+  outMin: new THREE.Vector3(),
+  outMax: new THREE.Vector3()
+};
+
+export function toThreeJsBox3(out: THREE.Box3, box: Box3, modelTransformation?: SectorModelTransformation): THREE.Box3 {
+  const { outMin, outMax } = toThreeJsBox3Vars;
+  out.set(toThreeVector3(outMin, box.min, modelTransformation), toThreeVector3(outMax, box.max, modelTransformation));
+  return out;
+}
+
+const fromThreeJsBox3Vars = {
+  min: vec3.create(),
+  max: vec3.create()
+};
+
+export function fromThreeJsBox3(box: THREE.Box3): Box3 {
+  const { min, max } = fromThreeJsBox3Vars;
+  fromThreeVector3(min, box.min);
+  fromThreeVector3(max, box.max);
+  return new Box3([min, max]);
 }
